put encoder in a box
